requirements.txt
absl-py==0.9.0
aiohttp==3.6.2
altgraph==0.16.1
astunparse==1.6.3
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
Automat==20.2.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
coverage==5.1
cryptography==2.9
cssselect==1.1.0
cycler==0.10.0
DateTime==4.3
decorator==4.4.2
elasticsearch==7.6.0
et-xmlfile==1.0.1
fabric==2.5.0
fake==0.8
fake-useragent==0.1.11
Faker==4.1.0
flake8==3.8.1
future==0.18.0
gast==0.3.3
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
hyperlink==19.0.0
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.6.0
incremental==17.5.0
invoke==1.4.1
jdcal==1.4.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lxml==4.5.0
Markdown==3.2.2
matplotlib==3.0.3
mccabe==0.6.1
more-itertools==8.3.0
mpmath==1.1.0
msgpack==1.0.0
multidict==4.7.5
mypy==0.770
mypy-extensions==0.4.3
networkx==2.4
numpy==1.18.4
oauthlib==3.1.0
opencv-python==4.2.0.34
openpyxl==2.6.4
opt-einsum==3.2.1
packaging==20.4
pandas==0.25.3
paramiko==2.7.1
parsel==1.5.2
pathlib2==2.3.5
pefile==2019.4.18
Pillow==7.1.2
pluggy==0.13.1
Protego==0.1.16
protobuf==3.12.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyflakes==2.2.0
Pygments==2.6.1
PyHamcrest==2.0.2
PyInstaller==3.5
PyNaCl==1.3.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pypiwin32==223
Pypubsub==4.0.3
PyQt5==5.15.2
PyQt5-sip==12.8.1
pytest==5.4.2
pytest-cov==2.8.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==225
pywin32-ctypes==0.2.0
PyYAML==5.3.1
pyyamlconfig==0.2.4
QtAwesome==1.0.1
QtPy==1.9.0
queuelib==1.5.0
requests==2.23.0
requests-oauthlib==1.3.0
robotframework==3.2.1
robotframework-ride==1.7.4.2
robotframework-ride-tortoisesvn==0.1.1
rsa==4.0
scikit-fuzzy==0.4.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.0.1
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.0.1
sympy==1.5.1
table==0.0.4
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
text-unidecode==1.3
Twisted==20.3.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
w3lib==1.21.0
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
wxPython==4.0.7.post2
xgboost==1.1.0
yagmail==0.11.224
yarl==1.4.2
zipp==1.2.0
zope.interface==4.6.0
